+++ src/__tests__/pages/api/dataSources/personnel.test.ts
@@ -22,6 +22,8 @@ const mockData = {
   lastModifiedAt: '2021-10-01',
 }
 
+const mockLastModifiedAt = '2021-10-01'
+
 describe('Personnel API', () => {
   describe('getUserData', () => {
     test('call getUserData with an id', async () => {
@@ -31,7 +33,7 @@ describe('Personnel API', () => {
       personnelAPI.baseURL = 'https://example.com'
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const mockPost = jest.spyOn(PersonnelAPI.prototype as any, 'post');
+      const mockPost = jest.spyOn(PersonnelAPI.prototype as any, 'post')
       mockPost.mockResolvedValue({ data: mockData })
 
       const result = await personnelAPI.getUserData('1234567890')
@@ -59,7 +61,7 @@ describe('Personnel API', () => {
       personnelAPI.baseURL = 'https://example.com'
 
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      const mockPost = jest.spyOn(PersonnelAPI.prototype as any, 'post');
+      const mockPost = jest.spyOn(PersonnelAPI.prototype as any, 'post')
       mockPost.mockResolvedValue({ data: mockData })
 
       const result = await personnelAPI.getGuardianDirectory()
@@ -76,6 +78,62 @@ describe('Personnel API', () => {
       await expect(personnelAPI.getGuardianDirectory()).rejects.toThrow(
         'No Personnel API URL found'
       )
+    })
+  })
+
+  describe('getLastModifiedAt', () => {
+    test('call getLastModifiedAt with an id', async () => {
+      const mockCache = {} as KeyValueCache
+      const personnelAPI = new PersonnelAPI({ cache: mockCache })
+
+      personnelAPI.baseURL = 'https://example.com'
+
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const mockPost = jest.spyOn(PersonnelAPI.prototype as any, 'post')
+      mockPost.mockResolvedValue({ data: mockLastModifiedAt })
+
+      const result = await personnelAPI.getLastModifiedAt()
+
+      expect(result).toEqual({ data: mockLastModifiedAt })
+    })
+
+    test('throw an error when there is no baseURL', async () => {
+      const mockCache = {} as KeyValueCache
+      const personnelAPI = new PersonnelAPI({ cache: mockCache })
+
+      personnelAPI.baseURL = ''
+
+      await expect(personnelAPI.getLastModifiedAt()).rejects.toThrow(
+        'No Personnel API URL found'
+      )
+    })
+  })
+
+  describe('searchGuardianDirectory', () => {
+    test('call searchGuardianDirectory with an id', async () => {
+      const mockCache = {} as KeyValueCache
+      const personnelAPI = new PersonnelAPI({ cache: mockCache })
+
+      personnelAPI.baseURL = 'https://example.com'
+
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      const mockPost = jest.spyOn(PersonnelAPI.prototype as any, 'post')
+      mockPost.mockResolvedValue({ data: mockData })
+
+      const result = await personnelAPI.searchGuardianDirectory('1234567890')
+
+      expect(result).toEqual({ data: mockData })
+    })
+
+    test('throw an error when there is no baseURL', async () => {
+      const mockCache = {} as KeyValueCache
+      const personnelAPI = new PersonnelAPI({ cache: mockCache })
+
+      personnelAPI.baseURL = ''
+
+      await expect(
+        personnelAPI.searchGuardianDirectory('1234567890')
+      ).rejects.toThrow('No Personnel API URL found')
     })
   })
 })