+++ src/components/SingleArticle/SingleArticle.tsx
@@ -48,6 +48,7 @@ export const SingleArticle = ({ article }: { article: ArticleRecord }) => {
           {videoTitle && <h2>{videoTitle}</h2>}
           {link && (
             <iframe
+              data-testid="embedVideo"
               title={videoTitle}
               width="420"
               height="315"